+++ CHANGELOG.md
@@ -101,6 +101,7 @@
    * ADDED: isochrone proper polygon support & pbf output for isochrone [#4575](https://github.com/valhalla/valhalla/pull/4575)
    * ADDED: return isotile grid as geotiff  [#4594](https://github.com/valhalla/valhalla/pull/4594)
    * ADDED: `ignore_non_vehicular_restrictions` parameter for truck costing [#4606](https://github.com/valhalla/valhalla/pull/4606)
+   * ADDED: log shortcuts which have the same start/end node [4632](https://github.com/valhalla/valhalla/pull/4632)
 
 ## Release Date: 2023-05-11 Valhalla 3.4.0
 * **Removed**